Refactor the structure of the Home page
- Remove Up and Down as special keys for SelectInput
- WIP: Add useFocusManager

+++ src/views/Home/index.tsx
@@ -47,9 +47,9 @@ const dateFormatter = new Intl.DateTimeFormat(
     [],
     {
         year: 'numeric',
-        month: 'long',
+        month: 'short',
         day: 'numeric',
-        weekday: 'long',
+        weekday: 'short',
     },
 );
